feat: restructuring catalogpage with outlet
